Export endpoint client in a new dispatch.test package
