+++ src/dispatch/test/__init__.py
@@ -1,0 +1,3 @@
+from .client import EndpointClient
+
+__all__ = ["EndpointClient"]
